Develop be

See merge request s12-webmobile2-sub1/S12P11A807!188

+++ backend/src/main/java/com/dawn/backend/domain/pin/entity/Pin.java
@@ -23,9 +23,9 @@ public class Pin extends BaseTimeEntity {
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	private Long pinId;
 
-	private Integer pinX;
+	private Float pinX;
 
-	private Integer pinY;
+	private Float pinY;
 
 	private String pinName;
 
@@ -35,8 +35,8 @@ public class Pin extends BaseTimeEntity {
 
 	@Builder
 	public Pin(
-		Integer pinX,
-		Integer pinY,
+		Float pinX,
+		Float pinY,
 		String pinName
 	) {
 		this.pinX = pinX;
